# ride-management-service/openapi.yaml
openapi: 3.0.3
info:
  title: RideLink Ride Management API
  version: v1
  description: Ride requests, driver assignment, lifecycle transitions and participant retrieval. Fare and payment are handled by a separate service.
servers:
  - url: http://localhost:8083
security:
  - bearerAuth: []
paths:
  /api/v1/rides:
    post:
      summary: Create a ride request
      description: PASSENGER only. New ride starts in REQUESTED state.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRideRequest'
      responses:
        '201':
          description: Ride created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Ride'
        '400': { description: Invalid pickup or destination }
        '401': { description: Missing or invalid token }
  /api/v1/rides/{id}/assign:
    post:
      summary: Assign the nearest eligible driver
      description: Requesting passenger or ADMIN. Driver Service returns candidates nearest-first within 25 km; empty result returns 409.
      parameters:
        - $ref: '#/components/parameters/RideId'
      responses:
        '200': { description: Ride assigned }
        '404': { description: Ride not found }
        '409': { description: Not REQUESTED or no eligible available drivers }
        '502': { description: Driver Service error }
        '503': { description: Driver Service unavailable }
  /api/v1/rides/{id}/accept:
    post:
      summary: Driver accepts assigned ride
      parameters: [{ $ref: '#/components/parameters/RideId' }]
      responses:
        '200': { description: Status becomes ACCEPTED }
        '403': { description: Caller is not the assigned driver }
        '409': { description: Invalid state transition }
  /api/v1/rides/{id}/start:
    post:
      summary: Assigned driver starts accepted ride
      parameters: [{ $ref: '#/components/parameters/RideId' }]
      responses:
        '200': { description: Status becomes IN_PROGRESS }
        '403': { description: Caller is not the assigned driver }
        '409': { description: Ride must be ACCEPTED }
  /api/v1/rides/{id}/complete:
    post:
      summary: Assigned driver completes ride
      parameters: [{ $ref: '#/components/parameters/RideId' }]
      responses:
        '200': { description: Status becomes COMPLETED }
        '403': { description: Caller is not the assigned driver }
        '409': { description: Ride must be IN_PROGRESS }
  /api/v1/rides/{id}/cancel:
    post:
      summary: Cancel an unstarted ride
      parameters: [{ $ref: '#/components/parameters/RideId' }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string, maxLength: 500 }
      responses:
        '200': { description: Status becomes CANCELLED }
        '403': { description: Caller is not a participant or ADMIN }
        '409': { description: Ride cannot be cancelled in its current state }
  /api/v1/rides/{id}:
    get:
      summary: Retrieve a ride by id
      parameters: [{ $ref: '#/components/parameters/RideId' }]
      responses:
        '200':
          description: Ride details
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Ride' }
        '403': { description: Caller is not a participant or ADMIN }
        '404': { description: Ride not found }
  /api/v1/rides/me:
    get:
      summary: Retrieve rides for the authenticated account
      responses:
        '200':
          description: Passenger rides, driver assigned rides, or all rides for ADMIN
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Ride' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    RideId:
      name: id
      in: path
      required: true
      schema: { type: string }
  schemas:
    Place:
      type: object
      required: [address, latitude, longitude]
      properties:
        address: { type: string, maxLength: 250 }
        latitude: { type: number, minimum: -90, maximum: 90 }
        longitude: { type: number, minimum: -180, maximum: 180 }
    CreateRideRequest:
      type: object
      required: [pickup, destination]
      properties:
        pickup: { $ref: '#/components/schemas/Place' }
        destination: { $ref: '#/components/schemas/Place' }
    Ride:
      type: object
      properties:
        id: { type: string }
        passengerAccountId: { type: string }
        assignedDriverAccountId: { type: string, nullable: true }
        pickup: { $ref: '#/components/schemas/Place' }
        destination: { $ref: '#/components/schemas/Place' }
        status:
          type: string
          enum: [REQUESTED, ASSIGNED, ACCEPTED, IN_PROGRESS, COMPLETED, CANCELLED]
        assignmentDistanceKm: { type: number, nullable: true }
        cancellationReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        assignedAt: { type: string, format: date-time, nullable: true }
        acceptedAt: { type: string, format: date-time, nullable: true }
        startedAt: { type: string, format: date-time, nullable: true }
        completedAt: { type: string, format: date-time, nullable: true }
        cancelledAt: { type: string, format: date-time, nullable: true }